Remove/move to other files tuples tests
